requirements.txt
# Install using pip install -r requirements.txt

# base requirements
asgiref
coverage
Django
sqlparse
# additional requirements (add new ones under here)


djangorestframework